Split the Ptcgo exporting/importing into their own implementations to seperate concerns since the importer isn't necessarily PTCGO specific
