+++ app/src/main/java/com/r0adkll/deckbuilder/arch/data/features/importer/repository/DefaultImporter.kt
@@ -1,29 +1,26 @@
-package com.r0adkll.deckbuilder.arch.data.features.ptcgo.repository
+package com.r0adkll.deckbuilder.arch.data.features.importer.repository
 
-import android.annotation.SuppressLint
+
 import com.r0adkll.deckbuilder.arch.data.AppPreferences
-import com.r0adkll.deckbuilder.arch.data.features.ptcgo.model.CardSpec
-import com.r0adkll.deckbuilder.arch.data.features.ptcgo.parser.DeckListParser
+import com.r0adkll.deckbuilder.arch.data.features.importer.model.CardSpec
+import com.r0adkll.deckbuilder.arch.data.features.importer.parser.DeckListParser
 import com.r0adkll.deckbuilder.arch.domain.features.cards.model.PokemonCard
 import com.r0adkll.deckbuilder.arch.domain.features.cards.repository.CardRepository
-import com.r0adkll.deckbuilder.arch.domain.features.ptcgo.repository.PTCGOConverter
-import io.pokemontcg.model.SuperType
+import com.r0adkll.deckbuilder.arch.domain.features.importer.repository.Importer
 import io.pokemontcg.model.Type
-import io.pokemontcg.model.Type.*
 import io.reactivex.Observable
 import timber.log.Timber
 import javax.inject.Inject
 
 
-class DefaultPTCGOConverter @Inject constructor(
+class DefaultImporter @Inject constructor(
         val repository: CardRepository,
         val preferences: AppPreferences
-) : PTCGOConverter {
+) : Importer {
 
     private val parser = DeckListParser()
 
 
-    @SuppressLint("CheckResult")
     override fun import(deckList: String): Observable<List<PokemonCard>> {
         return repository.getExpansions()
                 .onErrorReturnItem(emptyList())
@@ -85,62 +82,10 @@ class DefaultPTCGOConverter @Inject constructor(
                 }
     }
 
-    override fun export(cards: List<PokemonCard>, name: String): Observable<String> {
-        val pokemon = cards.filter { it.supertype == SuperType.POKEMON }
-        val trainers = cards.filter { it.supertype == SuperType.TRAINER }
-        val energy = cards.filter { it.supertype == SuperType.ENERGY }
-
-        val listBuilder = StringBuilder()
-        listBuilder
-                .appendln("****** $name ******")
-                .appendln()
-                .appendln("##Pokémon - ${pokemon.size}")
-                .appendln()
-
-        val groupedPokemon = pokemon.groupBy { it.id }
-        groupedPokemon.entries.forEach {
-            val count = it.value.size
-            val poke = it.value.first()
-            listBuilder.appendln("* $count ${poke.name} ${poke.expansion?.ptcgoCode ?: ""} ${poke.number}")
-        }
-
-        listBuilder
-                .appendln()
-                .appendln("##Trainer Cards - ${trainers.size}")
-                .appendln()
-
-        val groupedTrainers = trainers.groupBy { it.id }
-        groupedTrainers.entries.forEach {
-            val count = it.value.size
-            val trainer = it.value.first()
-            listBuilder.appendln("* $count ${trainer.name} ${trainer.expansion?.ptcgoCode ?: ""} ${trainer.number}")
-        }
-
-        listBuilder
-                .appendln()
-                .appendln("##Energy - ${energy.size}")
-                .appendln()
-
-        val groupedEnergy = energy.groupBy { it.id }
-        groupedEnergy.forEach {
-            val count = it.value.count()
-            val ener = it.value.first()
-            listBuilder.appendln("* $count ${ener.name} ${ener.expansion?.ptcgoCode ?: ""} ${ener.number}")
-        }
-
-        listBuilder
-                .appendln()
-                .appendln("Total Cards - ${cards.size}")
-                .appendln()
-                .appendln("****** Deck List Generated by DeckBox ******")
-
-        return Observable.just(listBuilder.toString())
-    }
-
 
     private fun filterEnergy(): (CardSpec) -> Boolean {
         return { spec ->
-            Type.VALUES.filter { it != COLORLESS && it != UNKNOWN && it != DRAGON}
+            Type.VALUES.filter { it != Type.COLORLESS && it != Type.UNKNOWN && it != Type.DRAGON }
                     .find {
                         spec.name.contains("${it.name} Energy", true) ||
                                 (spec.name.contains(it.name, true) && spec.set.contains("Energy", true))
@@ -151,7 +96,7 @@ class DefaultPTCGOConverter @Inject constructor(
 
     private fun mapEnergy(): (CardSpec) -> Pair<CardSpec, String>? {
         return { spec ->
-            val type = Type.VALUES.filter { it != COLORLESS && it != UNKNOWN && it != DRAGON }
+            val type = Type.VALUES.filter { it != Type.COLORLESS && it != Type.UNKNOWN && it != Type.DRAGON }
                     .find {
                         spec.name.contains("${it.name} Energy", true) ||
                                 (spec.name.contains(it.name, true) && spec.set.contains("Energy", true))
