+++ app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/exporter/tournament/TournamentExporter.kt
@@ -1,9 +1,9 @@
-package com.r0adkll.deckbuilder.arch.domain.features.tournament.exporter
+package com.r0adkll.deckbuilder.arch.domain.features.exporter.tournament
 
 
 import android.content.Context
 import com.r0adkll.deckbuilder.arch.domain.ExportTask
-import com.r0adkll.deckbuilder.arch.domain.features.tournament.model.PlayerInfo
+import com.r0adkll.deckbuilder.arch.domain.features.exporter.tournament.model.PlayerInfo
 import io.reactivex.Observable
 import java.io.File
