+++ app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/exporter/tournament/model/AgeDivision.kt
@@ -1,0 +1,8 @@
+package com.r0adkll.deckbuilder.arch.domain.features.exporter.tournament.model
+
+
+enum class AgeDivision {
+    JUNIOR,
+    SENIOR,
+    MASTERS
+}
