+++ app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/exporter/tournament/model/Format.kt
@@ -1,0 +1,7 @@
+package com.r0adkll.deckbuilder.arch.domain.features.exporter.tournament.model
+
+
+enum class Format {
+    STANDARD,
+    EXPANDED
+}
