--- app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/tournament/model/AgeDivision.kt
@@ -1,8 +1,0 @@
-package com.r0adkll.deckbuilder.arch.domain.features.tournament.model
-
-
-enum class AgeDivision {
-    JUNIOR,
-    SENIOR,
-    MASTERS
-}
