--- app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/tournament/model/Format.kt
@@ -1,7 +1,0 @@
-package com.r0adkll.deckbuilder.arch.domain.features.tournament.model
-
-
-enum class Format {
-    STANDARD,
-    EXPANDED
-}
